refactor: blur와 opacity 값 변경해주는 코드 재작성
- 기존 intepolateBlur 코드 네이밍 수정
  DOM의 실제 값을 바꾸는 코드의 이름이 모두 modify로 시작하므로 해당 키워드를 추가하여
  함수 이름을 modifyBlurFrameBackground 로 재작성
- 네이밍 변경에 따른 로직 수정
  기존 interpolateBlur 코드에서는 opacity와 blurPixels를 계산하고 dom에 적용까지 하였으나
  modifyBlurFrameBackground 함수에는 dom에 적용하는 코드만 사용, 계산을 핸들러에서 하여
  관심사분리
