+++ src/components/InfoWindow/Contents/BlurFrame/eventHandler/slideUpWindow.ts
@@ -1,7 +1,8 @@
-import interpolateBlur from '@components/InfoWindow/view/interpolateBlur';
+import modifyBlurAndOpacity from '@components/InfoWindow/view/modifyBlurFrameBackground';
+import { blurFrameTween, blurOpacityTween } from '@constants/Tween';
 import { PopUpHeightsType } from '@constants/popUpHeights';
 import { SlideUpWindowEvent } from '@libs/types/customEvents';
-import { calcProgressRatio } from '@libs/utils/calc';
+import { calcInterpolation, calcProgressRatio } from '@libs/utils/calc';
 
 export interface HandleSlideUpWindowForBlurEventProps {
   popUpHeights: { [key in PopUpHeightsType]: number };
@@ -17,7 +18,18 @@ export const handleSlideUpWindowForBlurFrame =
       end: topHeight,
     });
 
-    interpolateBlur({ ratio });
+    const blurPixels = calcInterpolation({
+      max: blurFrameTween.max,
+      min: blurFrameTween.min,
+      ratio,
+    });
+    const opacity = calcInterpolation({
+      max: blurOpacityTween.max,
+      min: blurOpacityTween.min,
+      ratio,
+    });
+
+    modifyBlurAndOpacity({ blurPixels, opacity });
 
     if (e.currentTop === middleHeight) {
       const target = e.target as HTMLDivElement;
